Pass libraries with LDLIBS instead of LDFLAGS
If libraries are passed in LDFLAGS, they are not picked up correctly
starting from Ubuntu 11.04 and later, as Ubuntu turned on the linker
option --no-copy-dt-needed-entries. This option is meant to improve
robustness of linking, but can cause link failures, if libraries are not
placed in the linker's command line such that they are after the file
that uses them. In automatic makefile rules, the value of LDLIBS is
placed after the files that are compiled, whereas LDFLAGS are placed
at the beginning.

For more information:

https://wiki.ubuntu.com/NattyNarwhal/ToolchainTransition