Do not call script to load the icc environment variables
The location of this script depends on the version of icc. If no
icc is installed or if the version of icc does not match the exact
version here, the build fails. Instead of calling this script from
the make files, it should be called from the users .bashrc file
or something similar.